@@ -31,7 +31,7 @@
     }
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=14.0.0"
   },
   "files": [
     "LICENSE",
