@@ -16,7 +16,7 @@
   },
   "main": "build/index.js",
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=14.0.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.11.2",
