@@ -3,9 +3,9 @@
   "version": "3.15.0",
   "sideEffects": false,
   "description": "React components for translations",
-  "main": "index.js",
+  "main": "./build/cjs/index.js",
   "module": "./build/esm/index.js",
-  "types": "./build/cjs/index.d.ts",
+  "types": "./build/index.d.ts",
   "author": {
     "name": "Tomáš Ehrlich",
     "email": "tomas.ehrlich@gmail.com"
@@ -29,7 +29,14 @@
     "url": "https://github.com/lingui/js-lingui/issues"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=14.0.0"
+  },
+  "exports": {
+    ".": {
+      "require": "./build/cjs/index.js",
+      "import": "./build/esm/index.js"
+    },
+    "./package.json": "./package.json"
   },
   "files": [
     "LICENSE",
