@@ -1,3 +1,3 @@
 const babelConfig = require("./scripts/build/babel.config")
 
-module.exports = babelConfig({ modules: true })
+module.exports = babelConfig(true)
