@@ -1,5 +1,1 @@
-if (process.env.NODE_ENV === "production") {
-  module.exports = require("./cjs/compile.production.min.js")
-} else {
-  module.exports = require("./cjs/compile.development.js")
-}
+module.exports = require("./cjs/compile.js")
