@@ -2,8 +2,13 @@ export {
   I18nProvider,
   useLingui,
   withI18n,
+} from "./I18nProvider"
+
+export type {
   I18nProviderProps,
   I18nContext,
   withI18nProps,
 } from "./I18nProvider"
-export { Trans, TransProps, TransRenderProps } from "./Trans"
+
+export { Trans } from "./Trans"
+export type { TransProps, TransRenderProps } from "./Trans"
